requirements.txt
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asyncio==3.4.3
asyncpg==0.29.0
bcrypt==4.1.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cache2==0.2.2
fastapi-cli==0.0.4
fastapi-versioning==0.10.0
flower==2.0.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.10.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
json-logging==1.3.0
jwt==1.3.1
kombu==5.3.7
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.6
packaging==24.1
passlib==1.7.4
pendulum==3.0.0
pillow==10.4.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.0
pydantic-settings==2.3.4
pydantic_core==2.20.0
Pygments==2.18.0
pytest==8.2.2
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
rich==13.7.1
rsa==4.9
sentry-sdk==2.12.0
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
time-machine==2.14.2
tornado==6.4.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
zope.event==5.0
zope.interface==6.4.post2
